+++ src/onegov.election_day/onegov/election_day/app.py
@@ -143,6 +143,7 @@ def micro_cache_anonymous_pages_tween_factory(app, handler):
         '/votes/.*',
         '/election/.*',
         '/elections/.*',
+        '/catalog.rdf',
     )
 
     cache_paths = re.compile(r'^({})$'.format('|'.join(cache_paths)))